Moved contact to pages
