+++ TuneOut/src/App.jsx
@@ -3,10 +3,9 @@ import 'bootstrap/dist/css/bootstrap.min.css';
 
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import SiteHeader from './components/SiteHeader';
-import Home from './Home';
-import About from './About';
-import Contact from './Contact.jsx';
-import Playlists from './Playlists';
+import Home from './pages/Home';
+import About from './pages/About';
+import Contact from './pages/Contact.jsx';
 import './App.css';
 
 function App() {
@@ -18,7 +17,6 @@ function App() {
           <Route path="/" element={<Home />} />
           <Route path="/about" element={<About />} />
           <Route path="/contact" element={<Contact />} />
-          <Route path="/playlists" element={<Playlists />} />
         </Routes>
       </div>
     </Router>
